fix: prevent lua file from being garbage collected
The idea is the same as buffers. Use `L_ref()` to prevent gc from
collecting something. In fact, `Buffer_Writer` is a special case of
`Lua_Stream_Data` in which we try to stream a lua data structure. The
stream procedure is responsible for casting the reference back to what
its supposed to be. File streams dont need to bother with that since we
call `os.to_stream()`. The point is to just stop the garbage collector